Remove build provided references and separate JUnit tests
