+++ rxjava-contrib/rxjava-swing/build.gradle
@@ -5,8 +5,8 @@ targetCompatibility = JavaVersion.VERSION_1_6
 
 dependencies {
     compile project(':rxjava-core')
-    provided 'junit:junit-dep:4.10'
-    provided 'org.mockito:mockito-core:1.8.5'
+    testCompile 'junit:junit-dep:4.10'
+    testCompile 'org.mockito:mockito-core:1.8.5'
 }
 
 javadoc {
